+++ libraries/bio/src/monomer-works/monomer-hover.ts
@@ -19,7 +19,7 @@ import {monomerSeqToMolfile} from './to-atomic-level-utils';
 import {MonomerHoverLink} from './utils';
 import {getMolHighlight} from './seq-to-molfile';
 import {MonomerMap} from './types';
-import {ISeqHelper} from '../utils/seq-helper';
+import {IHelmToMolfileConverter, ISeqHelper} from '../utils/seq-helper';
 
 export const MonomerHoverLinksTemp = 'MonomerHoverLinks';
 
@@ -31,26 +31,53 @@ export function addMonomerHoverLink(seqColTemp: any, resLink: MonomerHoverLink) 
   seqColTemp[MonomerHoverLinksTemp] = mhhList;
 }
 
-export function buildMonomerHoverLink(
+/**
+ * Builds a monomer hover link between sequence and molecule columns
+ * throughPOM specifies if the sequence is translated through seqhelper using POM (default is false)
+ */
+export async function buildMonomerHoverLink(
   seqCol: DG.Column<string>, molCol: DG.Column<string>,
-  monomerLib: IMonomerLibBase, seqHelper: ISeqHelper, rdKitModule: RDModule
-): MonomerHoverLink {
+  monomerLib: IMonomerLibBase, seqHelper: ISeqHelper, rdKitModule: RDModule, throughPOM: boolean = false
+): Promise<MonomerHoverLink> {
   const seqSH = seqHelper.getSeqHandler(seqCol);
-  const isNucleotideHelmSequence = seqSH.isHelm() && (seqSH.alphabet == ALPHABET.RNA || seqSH.alphabet == ALPHABET.DNA);
+  const isNucleotide = (seqSH.alphabet == ALPHABET.RNA || seqSH.alphabet == ALPHABET.DNA);
+  const isNucleotideHelmSequence = seqSH.isHelm() && isNucleotide;
+  // if the conversion goes through polymer object model, it is aware that DNA/RNA need sugars and phosphates, therefore, for fasta
+  // of length 15, it will return 45 monomers. so we need to handle this case separately
+  const isNucleotideNonHelmSequence = !seqSH.isHelm() && isNucleotide;
+
+  const getSeqMonomerCorrectedPosition = (pos?: number) => {
+    if (pos == undefined) return null;
+    // if its dna or rna, we need to correct position as it will contain sugar and phosphate
+    // toAtomicLevel returns 15 monomers for 15 nucleotides (r(n)p is one monmer), while helm converter returns 45
+    if (isNucleotideHelmSequence)
+      return throughPOM ? pos : Math.floor(pos / 3);
+    if (isNucleotideNonHelmSequence)
+      return throughPOM ? pos * 3 + 1 : pos;
+    return pos;
+  };
+
+  const helmToMolfileConverter: IHelmToMolfileConverter | null = throughPOM ? await seqHelper.getHelmToMolfileConverter(monomerLib) : null;
   function buildMonomerMap(seqCol: DG.Column<string>, tableRowIdx: number): MonomerMap {
     const seqSH = seqHelper.getSeqHandler(seqCol);
-    const seqSS = seqSH.getSplitted(tableRowIdx);
-    const biotype = seqSH.defaultBiotype;
-    const seqMList: ISeqMonomer[] = wu.count(0).take(seqSS.length)
-      .map((posIdx) => { return {position: posIdx, symbol: seqSS.getCanonical(posIdx), biotype: biotype} as ISeqMonomer; })
-      .toArray();
+    if (!throughPOM) {
+      const seqSS = seqSH.getSplitted(tableRowIdx);
+      const biotype = seqSH.defaultBiotype;
+      const seqMList: ISeqMonomer[] = wu.count(0).take(seqSS.length)
+        .map((posIdx) => { return {position: posIdx, symbol: seqSS.getCanonical(posIdx), biotype: biotype} as ISeqMonomer; })
+        .toArray();
 
-    const alphabet = seqSH.alphabet as ALPHABET;
-    const polymerType = alphabet == ALPHABET.RNA || alphabet == ALPHABET.DNA ? PolymerTypes.RNA : PolymerTypes.PEPTIDE;
-    const monomersDict = getMonomersDictFromLib([seqMList], polymerType, alphabet, monomerLib, rdKitModule);
-    // Call seq-to-molfile worker core directly
-    const molWM = monomerSeqToMolfile(seqMList, monomersDict, alphabet, polymerType);
-    return molWM.monomers;
+      const alphabet = seqSH.alphabet as ALPHABET;
+      const polymerType = alphabet == ALPHABET.RNA || alphabet == ALPHABET.DNA ? PolymerTypes.RNA : PolymerTypes.PEPTIDE;
+      const monomersDict = getMonomersDictFromLib([seqMList], polymerType, alphabet, monomerLib, rdKitModule);
+      // Call seq-to-molfile worker core directly
+      const molWM = monomerSeqToMolfile(seqMList, monomersDict, alphabet, polymerType);
+      return molWM.monomers;
+    } else {
+      const helm = seqSH.getHelm(tableRowIdx);
+      const molWM = seqHelper.helmToAtomicLevelSingle(helm, helmToMolfileConverter!, false, false); // skip mol beautification step. not needed.
+      return molWM.monomers;
+    }
   }
 
   const monomerMapLruCache = new LRUCache<string, MonomerMap>({max: 100});
@@ -86,19 +113,12 @@ export function buildMonomerHoverLink(
           setMonomerHover(null);
           return true;
         }
-        const getSeqMonomerCorrectedPosition = () => {
-          if (!seqMonomer) return null;
-          // if its dna or rna, we need to correct position as it will contain sugar and phosphate
-          if (isNucleotideHelmSequence)
-            return Math.floor(seqMonomer.position / 3);
-          return seqMonomer.position;
-        };
         setMonomerHover({
           gridCell: targetGridCell,
           dataFrameId: seqCol.dataFrame.id,
           gridRowIdx: gridRowIdx,
           seqColName: seqCol.name,
-          seqPosition: getSeqMonomerCorrectedPosition() ?? -1,
+          seqPosition: getSeqMonomerCorrectedPosition(seqMonomer?.position) ?? -1,
           getSubstruct: (): ISubstruct | undefined => { // Gets monomer highlight
             if (!seqMonomer || seqMonomer.symbol === '*')
               return undefined;
@@ -107,7 +127,7 @@ export function buildMonomerHoverLink(
             if (!molMonomerMap)
               return undefined;
 
-            const monomerMap = molMonomerMap.get(getSeqMonomerCorrectedPosition()!); // single monomer
+            const monomerMap = molMonomerMap.get(getSeqMonomerCorrectedPosition(seqMonomer?.position)!); // single monomer
             if (!monomerMap) return {atoms: [], bonds: [], highlightAtomColors: [], highlightBondColors: []};
 
             const res: ISubstruct = getMolHighlight([monomerMap], monomerLib);