+++ libraries/compute-utils/reactive-tree-driver/src/runtime/LinkControllers.ts
@@ -26,6 +26,7 @@ export class ControllerBase<T> {
   protected checkInput(name: string) {
     if (!this.inputsSet.has(name)) {
       const e = new Error(`Handler for Link ${this.id} is trying to set an unknown input ${name}`);
+      console.error(e);
       grok.shell.error(e.message);
       throw e;
     }
@@ -34,6 +35,7 @@ export class ControllerBase<T> {
   protected checkOutput(name: string) {
     if (!this.outputsSet.has(name)) {
       const e = new Error(`Handler for Link ${this.id} is trying to set an unknown output ${name}`);
+      console.error(e);
       grok.shell.error(e.message);
       throw e;
     }