+++ packages/Chem/src/tests/chemprop-tests.ts
@@ -36,7 +36,7 @@ category('chemprop', () => {
     const column = await fetchWrapper(() => applyModelChemprop(binBlob, DG.DataFrame.fromColumns([smilesColumn]).toCsv()));
         
     expect(column.length, 30);
-  }, {stressTest: true});
+  });
 });
 
 function getParameterValues() {