+++ packages/DiffStudio/src/app.ts
@@ -24,7 +24,6 @@ import {getIVP, getScriptLines, getScriptParams, IVP, Input, SCRIPTING,
   CONTROL_SEP, STAGE_COL_NAME, ARG_INPUT_KEYS, DEFAULT_SOLVER_SETTINGS} from './scripting-tools';
 
   import {CallbackAction, DEFAULT_OPTIONS} from './solver-tools';
-
 import {unusedFileName, getTableFromLastRows, getInputsTable, getLookupsInfo, hasNaN, getCategoryWidget,
   getReducedTable, closeWindows, getRecentModelsTable, getMyModelFiles, getEquationsFromFile,
   getMaxGraphsInFacetGridRow, removeTitle,
@@ -186,7 +185,7 @@ const completions = [
   {label: `${CONTROL_EXPR.INPUTS}: `, type: 'keyword', info: INFO.INPUS},
 ];
 
-/** Control expressions completion utilite */
+/** Control expressions completion utility */
 function contrCompletions(context: any) {
   const before = context.matchBefore(/[#]\w*/);
 
@@ -1124,7 +1123,7 @@ export class DiffStudio {
     }
   } // clearSolution
 
-  /** Return form with model inputs */
+  /** Generate model inputs */
   private async generateInputs(ivp: IVP): Promise<void> {
     /** Return options with respect to the model input specification */
     const getOptions = (name: string, modelInput: Input, modelBlock: string) => {
@@ -1426,6 +1425,7 @@ export class DiffStudio {
       const script = DG.Script.create(scriptText);
       await FittingView.fromEmpty(script, {
         inputsLookup: ivp.inputsLookup !== null ? ivp.inputsLookup : undefined,
+        ivp: ivp,
       });
     } catch (err) {
       this.processError(err);