+++ libraries/bio/CHANGELOG.md
@@ -1,5 +1,13 @@
 # bio changelog
 
+## 5.50.1 (2025-02-20)
+
+Remove helm dependency in toAtomicLevel for Linear sequences
+
+## 5.50.0 (2025-02-20)
+
+Support toAtomicLevel and POM conversion for all notations
+
 ## 5.49.1 (2025-02-17)
 
 Fix helm splitter for marginal cases